Introduce new util for attaching additional info to email
New util function which can be reusable and also the template for products suggestions in the email now is reusable
